Removed a few warnings

+++ samples/FrustumCullingInstancing/src/Main.cpp
@@ -153,9 +153,9 @@ int main(int argc, char** argv) {
         float angleY = 0.0f;
         if (x != 0) {
             if (x > 0) {
-                angleY = mouseSpeed * dt;
+                angleY = mouseSpeed * (float)dt;
             } else if (x < 0) {
-                angleY = -mouseSpeed * dt;
+                angleY = -mouseSpeed * (float)dt;
             }
 
             Matrix3x3 yaw;
@@ -167,9 +167,9 @@ int main(int argc, char** argv) {
         float angleX = 0.0f;
         if (y != 0) {
             if (y > 0) {
-                angleX = mouseSpeed * dt;
+                angleX = mouseSpeed * (float)dt;
             } else if (y < 0) {
-                angleX = -mouseSpeed * dt;
+                angleX = -mouseSpeed * (float)dt;
             }
             
             Matrix3x3 pitch;
